example-application/procs/laplacian-azure/requirements.txt
azure-functions
boto3
requests